refactor: split projects routes into public and admin routers
