+++ backend/app/main.py
@@ -11,7 +11,7 @@ from slowapi.errors import RateLimitExceeded
 from app.config import get_settings
 from app.limiter import limiter
 from app.logging import setup_logging
-from app.routers import auth, contact, profile, projects
+from app.routers import admin_projects, auth, contact, profile, projects
 
 settings = get_settings()
 
@@ -43,6 +43,7 @@ app.include_router(projects.router, prefix=settings.api_prefix)
 app.include_router(auth.router, prefix=settings.api_prefix)
 app.include_router(profile.router, prefix=settings.api_prefix)
 app.include_router(contact.router, prefix=settings.api_prefix)
+app.include_router(admin_projects.router, prefix=settings.api_prefix)
 
 
 @app.exception_handler(Exception)
